sol-tracker/requirements.txt
requests
python-telegram-bot>=20.0
APScheduler>=3.10
python-dotenv
python-dotenv
